perf(ffi): expand public C API for Nokogiri-compatible Ruby binding (TODO 138)
The Ruby FFI team reported that the public C API was "too minimal"
for a Nokogiri-style binding. Added the missing pieces:

New functions in taurus.h:
- taurus_document_encoding(doc)        — was implemented but not in header
- taurus_element_attribute_count(elem)  — was in internal header, no TAURUS_API
- taurus_element_attribute_name_at(elem, i)  — new
- taurus_element_attribute_value_at(elem, i)  — new
- taurus_element_namespace_count(elem)  — new (walks attribute list)
- taurus_status_string(status)           — new (human-readable errors)

Implementation in:
- src/taurus/dom/element_query.c — attribute index accessors, namespace count
- src/taurus/error.c              — status string

Internal change:
- dom/element.h: taurus_element_attribute_count return type uint8_t -> size_t
  (FFI bindings need the full 0-2^64 range, not 0-255)

Symbol export check (scripts/check_shared_exports.sh) now requires
all new FFI helpers so the shared library export is verified.

Specs in test_header_hygiene.cpp:
- DocumentEncodingAccessor
- AttributeIndexAccess
- NamespaceCount
- StatusString

373/373 pass. ASAN clean. Symbol export verified.

+++ src/taurus/dom/element.h
@@ -361,8 +361,8 @@ struct taurus_attribute* taurus_element_get_first_attribute(TaurusElement elem);
 /* Set first attribute in element */
 void taurus_element_set_first_attribute(TaurusElement elem, struct taurus_attribute* attr);
 
-/* Get attribute count */
-uint8_t taurus_element_attribute_count(TaurusElement elem);
+/* Get attribute count (size_t for public API compatibility — TODO 138) */
+size_t taurus_element_attribute_count(TaurusElement elem);
 
 /* Get attribute by index */
 struct taurus_attribute* taurus_element_get_attribute_by_index(TaurusElement elem, uint8_t index);
